stable/swm-sync-jira: automatically assign only kernels part of cranky rotation
kernel-versions/<cycle>/info/reviewers.json file contains also the kernels
that do not have a reviewer (yet) with the value marked as null.
Swm will know then if the kernel has to be assigned to the cycle lead.
Any kernel that's not part of cranky rotation should not be automatically
assigned to anyone by swm, that's the responsability of the cranker.

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>