annotations-tools: do not try to auto-detect JSON format
The script tries to automatically detect the format of the annotations
file (legacy or pure JSON).

However, if the legacy file contains a syntax error, the script will try
to parse the file as JSON (incorrectly) triggering another obscure
error, complaining that the JSON format is invalid.

This hides the capability to detect syntax errors properly in the
annotations file.

To prevent this, introduce a new option --json. In this way the caller
has complete control to decide which format needs to be used to parse
the annotations file.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>