cranky: add check-fips command
The fips-checks script currently lives under the debian/scripts/misc/
directory in the Jammy, Mantic and Noble kernel trees. In an attempt to
get this type of script out of these trees, move fips-checks to be a
cranky command that is to be used as a step in `cranky close` for the
FIPS kernels and rename it to check-fips.

Signed-off-by: Magali Lemes <magali.lemes@canonical.com>