cranky/cmd.d/update-dkms-versions: Early return for xenial and trusty
This should not be used for xenial and trusty derivatives because the
migration was not fully deployed.
For other series, this command fails if update-dkms-versions script is
still in the tree, but for xenial and trusty, there is no such script.
Therefore this command will move debian/dkms-versions to
debian.<derivative>/dkms-versions. This is incomplete, because the build
rules/script still contain the old reference for xenial and trusty and
people may not realize they need to change that.
As a temporary fix, disable this for xenial and trusty.

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>