wfl/wft/kernel_snap.py: relax branch matching
The branch reported in the test-observer results is the _current_ branch
and not the tested one.  We therefore cannot use that to match results.
For V2 recipes we are able to and do check the revisions tested match
those we have.  For V1 ones we will have to assume these are highly
likely the right results.  Drop the branch selector.

Signed-off-by: Andy Whitcroft <apw@canonical.com>