chroot-setup: work-around quits when sourced under exit-immediately mode
`let ...` ensures the following statement is evaluated under arithmetic
mode, and when the statement is assigned to zero, bash interprets it as
a false return code, which triggers an unexpected error and quit the
shell when exit-immediately mode is on. This stops the reuse of the
chroot definitions conf in a few restricted cases.

Signed-off-by: You-Sheng Yang <vicamo.yang@canonical.com>