cranky/link-tb: Replace datetime.utcnow by datetime.now(UTC)
As utc now is deprecated, move to the latest api to silence this warning:

DeprecationWarning: datetime.datetime.utcnow() is deprecated and scheduled for removal in a future version. Use
timezone-aware objects to represent datetimes in UTC: datetime.datetime.now(datetime.UTC).

As this new interface is not yet available everywhere (including Jammy), support
previous version as a fallback.

Signed-off-by: Thibault Ferrante <thibault.ferrante@canonical.com>
Acked-by: Magali Lemes <magali.lemes@canonical.com>
Signed-off-by: Thibault Ferrante <thibault.ferrante@canonical.com>