cranky/cmds.d/rebase-selector: Include getabis script to ignore_files
debian/scripts/misc/getabis will be removed everywhere in this cycle,
but, in the process, many derivatives will have conflicts during rebase
because it is deleted in the parent, and any extra changes for this script
on top will fail.
This change will make the transition smoother, people won't have to run
git rebase --skip.

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>