stable/create-cycle-sprint: Include cycle-owner as param
The cycle owner is needed for 2 things:
1. assign the cycle task (otherwise they have to be manually assigned in
   the jira board)
2. store this info in sru_cycle.yaml file. This is later used to assign
   kernels to the owner for peer-reviewing in case there's nobody
   around.

Note: The id_to_email map is temporary. We have a repo kernel_users but
it requires SSO authentification. It will be moved under 10. subnet
(under the VPN) and after that I will remove this hardcoded map.

While at it, also add 'owner' field to 'Sync-forward kernels' task so it is
automatically assigned to Andy.

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>