ancillary-bot/ancillary-{config,lookup}: handle kernels with ancillaries but no private-build
When we have ancillaries but no private-build route we will not consider
the PPA as needing scans.  This fails when the only kernels in a PPA
have signed and generate packages but no LRM.  Add those without an
adjunct reference.

Handle missing adjunct reference if we attempt to use it.

Signed-off-by: Andy Whitcroft <apw@canonical.com>