docs: use cranky tags and cranky update-dependents
The current guide for cranking the kernel does not make use of the
`cranky update-dependents` and `cranky tags` commands that automate
going into the correct dependents to update and tag them, instead of
doing it manually.

Update the guide to use the smart commands, and keep the manual way
explained in case it is needed.

Since the `cranky tags` also tags the `linux-main`, it has to be moved
after the `cranky update-dependents` step so the command tags both
`linux-main` and its dependents.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>
Acked-by: Thibault Ferrante <thibault.ferrante@canonical.com>
Signed-off-by: Thibault Ferrante <thibault.ferrante@canonical.com>