docs: Add section on Jira ticket for peer-review
The person that prepared the kernel is responsible for changing the
status to Kernel Peer Review.
If the kernel was manually assigned by a colleague, they are also
responsible to change the assignee. If it was automatically assigned by
swm as part of cranky rotation, no need to manually change the assignee,
this is done automatically by shiv when the status changes to Kernel
Peer Review.

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>
Acked-by: Magali Lemes <magali.lemes@canonical.com>
Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>