cranky: mark test_cycle_ref tests as too slow
Unit tests should be very fast, especially when integrating
them in a pre-commit hook. For some reason, these tests are
taking aroud one second each on a beefy machine.

If the tests are doing network calls or other busywork, they
should not be unit tests but integration tests that are run
at a lower frequency.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>