cranky/cranky.py: Introduce new python cli
The existing cranky.py entry point is currently only used by the snap
and simply calls the bash entry point which then calls the relevant
subcommand. Yikes :-) As a first step towards a python-only cranky
toolset, make cranky.py a click-based cli with native click commands
for the individual existing subcommands. For the time being, these
click commands are still wrappers that call the 'old' subcommands in
'cmds.d'. The goal is to slowly convert them to proper click commands...

Introduce a new library crankylib that provides the new bits and pieces
which are:
  1) a new cranky class that is available to all subcommand and can hold
     and provide common things like configs and helpers and such.
     Currently it only provides a call() method that invokes the old
     subcommand.
  2) cli helper functions that simplify the construction of click
     commands. Most of this can be removed once all subcommands have been
     converted to native click commands.

Note that this change is totally transparent to existing users of the
repo and should not impact anybody other than snap users (of which there
are not many :-)).

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>