CONTRIBUTING: explain how to make changes to Gitea files
This is mostly a copy-paste of the (long) discussion we had to
understand how one could make changes to the pr_check.yml file
as a simple commiter with no other rights.

We agreed that this process is convoluted and should be improved.
Document the current process while we figure out something better.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>
Acked-by: Thibault Ferrante <thibault.ferrante@canonical.com>
Signed-off-by: Agathe Porte <agathe.porte@canonical.com>