cranky: crl/config: introduce ConfigPath
The configuration path has a notion of obsolete path that was
burried behind the `warn` flag. Extract that behavior into a
simple class with the more explicit `is_obsolete` attribute.

Make ConfigPath.default() the only source of truth to avoid
duplication of the path in the warning string shown if the
path is obsolete.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>