cranky: crl/config: introduce Config.default
In order to split the constructor into multiple independent
funtions, we need to have a default contructor that does not
require crazy arguments first.

Make Config a dataclass, use its default constructor, and
introduce a separate Config.default() constructor that
will contain the previous implementation to be splitted in
the next commits.

The `warn = True` assignments were useless as overriden by
the obsolete options check block. They were removed, and
the checks are now in a dedicated method that is called after
the default ctor.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>