cranky/cranky.py: Add all subcommands from 'commands' directory
Programmatically add all subcommands from the 'commands' directory by
reading the files list, inferring the callback names from the file
names and adding them to the top cli group. It is required that the
file basenames and the callback function names are identically.

This is basically identically to doing it manually for each command
like: so
  from commands import <xyz>
  cli.cranky_cli.add_command(<xyz>.<xyz>)

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>