cranky: crl/config: do not use typing.Self
This is a bit too modern and does not work on jammy because it was
introduced in Python 3.11. Leave a TODO note about it, since we
can hope to drop jammy support once Noble reaches maturity with its
first dot release and users migrate to it.

We could have introduced the `typing_extension` PyPI dependency
to backport typing features to older Python interpreters, but
introducing a dependency for such a small inconvenience seems overkill.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>