cranky/cmds.d/update-dkms-versions: add support for absolute references
We have a big problem with dkms-versions for Nvidia becoming out of sync
between cycles, particularly when creating new cycles in parallel with
the Nvidia versions changing.  In order to avoid this we desire the
Nvidia versions to be consistent in all cycles.  Add support for
absolute references outside of the cycle.  We will use this to implement
a /common/dkms-versions containing the Nvidia data.

Signed-off-by: Andy Whitcroft <apw@canonical.com>