mainline-builds: cod-mainline-core -- Fall back to default gcc version
Newer kernels specify an explicit gcc version to use, which might not
be available in the build chroot. In this case, fall back to using the
installed default gcc version.

Side note: It's not pretty to modify a build rule file for that. We
should be able to do that in the hooks file but that file is included
after the fact...

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Acked-by: mehmetbasaran <mehmet.basaran@canonical.com>
Acked-by: Timo Aaltonen <timo.aaltonen@canonical.com>
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>