mainline-build/cod-incoming-publish: Fix tag push of retriggered builds
If a mainline build is retriggered, a new tag (cod/mainline/<version>) and
crack.bundle are generated. In this case we need to force-update any already
existing tag, both locally and remotely.

Example: A mainline build of v6.9.4 was retrigged which resulted in:

cod-incoming-publish: mainline/v6.9.4: publishing crack.bundle ...
error: tag 'refs/tags/cod/mainline/v6.9.4' not found.
From /home/ubuntu/mainline-builds/xfer/mainline/v6.9.4/crack.bundle
 ! [rejected]                  cod/mainline/v6.9.4 -> cod/mainline/v6.9.4  (would clobber existing tag)

The result of this issue is that the advertised commit sha (of the tag)
doesn't exist in the mainline-crack LP repo, the remote tag is not updated
and subsequently the builders build the packages from the old tag.

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>