cranky/crl/git.py: get_current_tag when the tag does not point to HEAD
For cases where you prepare a kernel but you leave some commits on top
in your branch for the next release (usually it's for respins but not
only), the latest tag does not point to HEAD, but to some older commit
in the history.
cranky push-refs will not push the tag in this case.
This is fixed by getting the latest tag present.
This method is being used by cranky push-refs, therefore I changed the
implementation to keep it simple. If we ever to get a tag that's
pointing to HEAD, then another method can be created.

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>