ktl/sru_spins.py: new accessor library for spin/handle information
Adds a new accessor library to obtain tracker and version information
for SRU spins.  With a spin (or cycle) and a handle we can obtain an
information record for that kernel if it was spun.  This data includes
the tracker number to allow finding the full details and the per package
version data to allow checkouts of that spin.

Signed-off-by: Andy Whitcroft <apw@canonical.com>