swm: wft/promoter -- also require GPU-specific pointers
The CPC team will publish current-gpu-driver-$GPU-* in addition to
current-driver-* pointer files. The code is there but we are waiting
for Google to tell us that we can enable GPU-specific drivers.

These pointer files are used for the specific $GPU attached to the
system and they will usually contain a driver version higher than the
default specified in current-driver-*.

For example, if the GPU identified by NVIDIA_TESLA_B200 will require 550
driver version, the kernel-series.yaml will be updated with the patch
below to make the promoter.py wait for all required files.

  diff --git a/info/kernel-series.yaml b/info/kernel-series.yaml
  index bc92210..0e4ec89 100644
  --- a/info/kernel-series.yaml
  +++ b/info/kernel-series.yaml
  @@ -4776,6 +4776,9 @@ defaults:
             - '1.29'
           nvidia-releases:
             - 535
  +          - 550
  +        gpu-overrides:
  +          - NVIDIA_TESLA_B200
         testing:
           flavours:
             gke:

Signed-off-by: Tomáš Virtus <tomas.virtus@canonical.com>