stable/swm-jira-sync: Cleanup peer-reviewer automatic assignment
This patch removes the nested if statements under the Peer Reviewer
state.
Moreover, it changes the state to Kernel Building if the kernel has been
uploaded by the peer reviewer.
Once the kernel is uploaded, the owner is assigned back to the jira
card. This is done no matter the initial jira card state because once in
a while we have a separate squad for preparation and kernels have to be
assigned back to the owner. This is rare, mostly this is needed if
someone has peer reviewed the kernel.

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>