ancillary-bot/ancillary-{lookup,builder}: use kernel-series adjunct marker
Currently builder assumes all ancillaries are adjunct except for
linux-generate.  ancillary-lookup knows which are adjunct and uses this
to select an appropriate PPA to scan so it makes sense for it to expose
this information to ancillary-builder.

Add the selected destination (primary or adjunct) reference to the
result rows.  Also add the visibility (Public or Private) for that
destination to allow ancillary-builder to perform last ditch checks to
prevent linux-restricted-* being uploaded to a public destination.

Signed-off-by: Andy Whitcroft <apw@canonical.com>