ancillary-bot/ancillary-{lookup,builder}: record completed uploads
When we complete an upload record it and pass the completed list back to
ancillary-lookup to allow it to short-circuit processing in a similar
manner to the ignore list.  For now we are not clearing out dead entries
from the ignore lists, so keep the uploaded list separate so we can
drop it and let it be reconstructed.

Signed-off-by: Andy Whitcroft <apw@canonical.com>