stable/wfl/package.py: handle bug list correctly when aleady moved to Updates
When a bug moves to updates its version will match updates and we cannot
use the Changes: to augment the bug list.  Currently when this happens
we end up with an empty bug list, leading to a full request for
validation in launchpad.  Detect this and fallback to the base
Launchpad-Bugs-Fixed: data.

Signed-off-by: Andy Whitcroft <apw@canonical.com>