cve-tools/cves-kernel-versions-prepare: consider top layer -updates only for -release
We use layered -updates pockets to allow us to track the cross over from
primary archives to ESM and from ESM to ESM-legacy.  However security's
definition of `not-affected` considers the start of that category as
from the first release into that pocket.

In order to deliver this semantic we need to consider kernels with
multi-entry "updates" route specially.  Specifically when we have
multiple layers for "updates we only consider the top layer of this for
"first publication" purposes.

Signed-off-by: Andy Whitcroft <apw@canonical.com>