cve-tools/cve-rejuvenate: handle flipping devel -> oracular -> devel as needed
When rejuvenating retired CVEs we wish to map devel to an appropriate
release, fix this for the current development release.  Also add
handling for mapping the current development release to devel in active
CVEs.  This will allow us to rejuvenate a CVE in retired/ realise we
need it moved back to active/ and then rejuvenate it again.

Signed-off-by: Andy Whitcroft <apw@canonical.com>