cranky/crl/tests/test_handle.py: test_directory_source_noconfig -- this test is now invalid
This (badly documented) test  was intended to test the case where the
cranky user has not yet configured their cranky setup.  However we have
changed the cranky configuration handling to add default values.  This
means that this test can no longer pass as there is no such thing as
having no config.  This has been hidden by other bugs until recently.
This is particularly problematic as the default configuration in cranky
(using ~canonical) matches the common setup for local user testing.  So
it will fail in gitea and will work when tested locally for those users.

Drop this test as it simply no longer makes sense.

Signed-off-by: Andy Whitcroft <apw@canonical.com>