cve-tools/integrate-autotriage: separate out trivial changes to fips and EOL series
We do not really need to review updates to needed for fips/ and fips-preview/
as these are not updated reactivly they are updated based on the
recertification schedule.  We also do not need to worry about things
which are updating to EOL closed series, though we should be pulling
them into our matrix as generally active.

Signed-off-by: Andy Whitcroft <apw@canonical.com>