kernel-stable-dashboard.mako: use CSS tooltips to display readme data
We need a rich format to display the Cycle README data extracted from
Jira.  We use things like pre-formatted text and style changes.  Also
traditional tooltips are not accessible on mobile devices where hover is
not a thing.  Use using CSS based tooltips which are full <span> objects
as this supports full HTML.  They also can be linked up to clicks on a
new icon.

Signed-off-by: Andy Whitcroft <apw@canonical.com>