ktl/sru_spins.py: switch up to record by handle
All consumers of the spin data are accessing it via handle/spin or
handle/cycle lookups.  For cranky use we would like to get a list of
valid spins per handle.  If we reshape the data to be stored by handle
we will make this much simpler and cheaper to calculate.  This also will
reduce the size of the data we need to fetch in the average case.

Signed-off-by: Andy Whitcroft <apw@canonical.com>