bwheeler/deprecate-cranky-cheatsheet (#757)
The cranky documentation has a "cheatsheet" feature that pulls various info (typically marked with a `<!--cheatsheet-->` flag) from the markdown files, and can optionally turn it into HTML.

People don't seem to use this. Additionally, it requires some positional ordering of the documentation that is not ideal for coherent and concise doc-writing (I can provide details if necessary).

As part of a goal to clean up the docs, I intend to deprecate this cheatsheet functionality. It will allow us to write more clear and concise docs, hopefully to the point at which the script won't be necessary anymore!

Reviewed-on: https://kernel.ubuntu.com/gitea/kernel/kteam-tools/pulls/757
Acked-by: Kevin Becker <kevin.becker@canonical.com>