ktl: fix shebangs and +x flags
Some scripts have the +x flag and a shebang, but contain no code to
run in the script body. Remove the executable flag and the shebang
from these files to signal the intent that they should not be executed
directly.

Other files have a shebang but are not executable so they are useless.
Remove these too.

Finally some scripts have a shebang and a main function, but are not
marked as executable. Add the +x flag to these files to be able to
see easily which files can be executed.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>