snaps/snap-recipe-create: add support for updating existing snap recipies
Add support for detecting when we have already made the relevant snap
recipies.  When they already exist update any parameters which we know
how to change.  Dump the new recipe specification in a more human
readable form.  Finally, document this new use.

NOTE: this is also added to the proceedures for series-opening.

Signed-off-by: Andy Whitcroft <apw@canonical.com>