stable/wfl/package.py: fix stable pocket ordering
scan_pockets for a series is assumed to list the valid pockets for that
series.  This basically selects for updates/security or release
depending on the series.  This list is assumed to be ordered in workflow
order and is consumed to determine pocket deltas.  The current
scan_pocket for the stable series has the Updates and Security pockets
in the wrong order.  Correct this.

Signed-off-by: Andy Whitcroft <apw@canonical.com>