dashboard/ktml/matching.py: switch to direct KernelSeries lookups to populate cache
All callers of really expect the list of handles to be contemporaneous.
We risk reusing a historical dump though the use of a file to cache the
data.  Now that we have a data-level cache we can simply access
KernelSeries() directly and cache that result for the length of the job.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Thibault Ferrante <thibault.ferrante@canonical.com>
Signed-off-by: Thibault Ferrante <thibault.ferrante@canonical.com>