stable/wfl/package.py: record linux-image-flavour by arch for meta
Stakeholders would like to be able to tell which linux-image-flavours
are expressed by a particular kernel to ensure they are all tested
before signoff.  Add a new `swm-properties` entry `flavours` which lists
the flavours and their architectures.  These are generated based on the
binary packages expressed by the linux-meta package in the kernel.

Signed-off-by: Andy Whitcroft <apw@canonical.com>