lpltk: Move it to a top-level directory
There are 3 similar local libraries: ktl, pika and lpltk. We want to
organize them all identically, meaning they all live in a top-level
directory <foo> and there is a symlink libs/<foo> -> ../<foo>.

So move the 'lpltk' directory to the top, add the symlink and remove the
now empty/useless 'py3' top-level directory.

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>