ktl/kernel_series.py: fix local mode and data_location
Local data mode has been broken for the "tip" for a long time.  That
data moved to join its peers in kernel-versions long ago.  This means
that at its core all of this mode is broken.  Additionally we have
utterly confusing semantics for selecting local mode local selection
(use_local on lookup interfaces).  Finally how "local" data selectors
are passed on to the nested SigningConfig is broken as we can only
pass on `use_local`.

Move to a model when fundamentally we supply `data_location` to the
KernelSeriesCache level in order to set the base URL from which data is
fetched.  This may either be directly supplied as is done in the test
suite or inferred from the KERNEL_SERIES_USE configuration.  This
`data_location` is a URL which points to the directory in which
`kernel-series.yaml` is located (cycles are assumed in the directory
above when consuming raw-yaml data forms).  This allows the same data
location to be passed on directly to SigningConfig.  When we select
`local` data this is now in a single consistent location in
`../kernel-versions/` relative to the library.  This allows us to simply
symlink a local kernel-versions table into the top-level.

Signed-off-by: Andy Whitcroft <apw@canonical.com>