kernel_series: fix broken executable flag
The file cannot be executed directly because of relative imports.
However the main can still be executed by calling
`python3 -m ktl.kernel_series` in the parent folder.

Remove the executable flag, the shebang, and add a comment before
the ifmain block to help users run this file if needed.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>