ktl: drop ubuntu.py and scripts that use it
The Ubuntu class have been using an urlopen on a 404 link for quite some
time. None of the scripts are handling the generated
urllib.error.HTTPError: HTTP Error 404: Not Found error that is raised
when the class is instanciated, which is a good indication of unused
code.

Remove the failing ubuntu.py file and all other scripts that depend on
it because they have been failing for a long time (most are python2
anyways).

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Agathe Porte <agathe.porte@canonical.com>