cranky: Introduce cranky format-kernel-series
This is useful for formatting kernel-series.yaml in the kernel-versions
repo. It only formats the kernel-series.info files modified.

The formatting logic is implemented in ktl and the cranky command uses
this.

This tool was originally implemented in the kernel-versions repo because the
actual files are there. But decoupling the command from the data is the
way forward and the scripts from kernel-versions will be moved to
kteam-tools. One important reason is that we have duplication in both
repos or even different implementation for the same thing. Moving the
scripts from kernel-versions reduces this overhead.

Note: Ideally, this should not be a cranky command, because cranky is supposed to
be used only for kernel preparation. In the future a separate group of
commands will be created and this will be moved there. Until then, it is
a cranky command.

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>