Remove kernel-sru-workflow
This dashboard has been unpopulated for Noble, Oracular and Plucky.
The last recorded cycle displayed in the online version is 2023.09.14:

https://kernel.ubuntu.com/reports/kernel-sru-workflow/

This means that the dashboard has not been used since 2023.09.

Instead of refactoring it to drop the lpltk dependency and use
launchpadlib, just drop it.

Note that the stable/dashboards/do-sru-workflow script is not called
anywhere in the repository, which means that something external is
calling it. Depending on how it is done ("./do-*" ?) then the host
that generates the dashboards might need to be updated.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>