cranky: Introduce cranky distribute command
This will reassign kernels to people. The algorithm used is very basic.
Based on a list of kernels and a map of owners with their maximum number
of kernels, it will go through the list of kernels and assign them to
the owners. The list of kernels is always shifted by an index so that
every cycle, kernels are assigned to (somewhat) different owners. The
index is calculated based on how many sru_cycles have been done after
2024.01.08. This algorithm does not ensure that kernels are always
assigned to someone else every cycle.

The input data for this command is from your local kernel-versions repo.

It exports two steps:
- step 1:
Reassign (easy) kernels to people working on upload rights. They will get
a peer-reviewer assigned as well. The input files for this are
flexible_kernels_step_1, flexible_owners_step_1 and peer_revieweres from
<kernel-versions>/info/cranky_distribute

- step 2:
Reassign kernels for people who volunteered to help the official owners.
The input files are flexible_kernels_step_2 and flexible_owners_step_2
from the same location as step1.
These are populated by managers right before the cycle starts, after the
first round of cranky distribute was run. This should be used only in cases
where the official owners are temporarily busy with something else for a
cycle and there are volunters that can take over.
DO NOT use the second step as permanent reassigment. For this, please
just reassign kernels in global info/kernel-series.yaml file.

Both steps result in the same thing: changing owners and writing in the
kernel_series.yaml file for that specific cycle.

NOTE: Same as cranky format-kernel-series, this should not be a cranky
command, because cranky is supposed to be used only for kernel preparation.
In the future a separate group of commands will be created and this will
be moved there. Until then, it is a cranky command.

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>