stable/wfl/snap.py: add snap-pkg into package to avoid naming collision
Compose a new "package" entry by appending a "-" and "snap-pkg", to
avoid naming collision on deb and snap testing.

One good example is the pi-kernel, where the "package" is linux-raspi.
If we are using linux-raspi for both deb and snap testing, the test
result will be listed under the same kernel.

As we are not using package for snap testing, it is ok to modify it
for the snap test request.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>