info: Remove repositories.txt
In the old days, this was the definition of repos to be synced to the
builders, to the queue manager (kernel.u.c aka wani) and between wani
and Launchpad. This all changed with the move to PS5 and the
decommission of wani.

The syncing between new kernel.u.c and Launchpad is now handled by
mainline-build/cod-update-virgin. The syncing of repos to the builders
is handled by kernel-team-accounts/update-repos2.sh. Neither of them is
using info/repositories.txt anymore so remove it.

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>