stable/wfl/wft/regression_testing.py: allow snap kernels to pickup the right test result
We are emitting package-snappkg as source package name for snap kernels
to avoid naming collision with deb kernels. (We canot use / in the source
pacakge name otherwise it will be a pain to handle filename issues.)

Therefore replacing the ' / ' in the target, allow snap kernels to pickup
the right test result from summarized-results.json

As we only have boot test enabled for snap kernels, this change is made
exclusively in BootTesting class.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>