cranky: distribute: fix flexible kernels list consumption
The distribute_kernels() was using the provided kernel list object
directly and performing operations such as pop() on it. This was
changing the original object, therefore the list would be empty on a
second call of the function. This was breaking for example the
assignment of the peer-reviews, which needs to be done by walking the
flexible kernel list a second time.

Fix this issue by making a copy of the list object before using it.

Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>